Created Profile UI for Navbar
